refactor(imports): reorganize component import paths for consistency
Updated import paths across multiple components to enhance code organization and maintain consistency. This includes moving diagnostic and shared components to their respective directories, ensuring a clearer project structure.
