+++ src/app/canvas/page.tsx
@@ -27,12 +27,12 @@ import { RoomContext } from '@livekit/components-react';
 import { useRouter } from 'next/navigation';
 import { useAuth } from '@/hooks/use-auth';
 import { ToolDispatcher } from '@/components/TO BE REFACTORED/tool-dispatcher';
-import { SystemRegistrySync } from '@/components/ui/system-registry-sync';
+import { SystemRegistrySync } from '@/components/ui/diagnostics/system-registry-sync';
 import { initializeMCPBridge } from '@/lib/mcp-bridge';
-import { AgentCapabilitiesBridge } from '@/components/ui/agent-capabilities-bridge';
+import { AgentCapabilitiesBridge } from '@/components/ui/integrations/agent-capabilities-bridge';
 import { LiveKitStateBridge } from '@/lib/livekit/livekit-state-bridge';
 import LiveKitDebugConsole from '@/components/LiveKitDebugConsole';
-import DevEventFeed from '@/components/ui/dev-event-feed';
+import DevEventFeed from '@/components/ui/diagnostics/dev-event-feed';
 // TODO: Investigate best way to "go back" to CanvasSpace once we have a better way to handle adding/updating/managing the state of multiple components on the canvas simultaneously
 
 // Suppress development warnings for cleaner console
