+++ src/components/ui/messaging/message-thread-collapsible.tsx
@@ -11,7 +11,7 @@ import { type VariantProps } from 'class-variance-authority';
 import { useRoomContext } from '@livekit/components-react';
 import { RoomEvent } from 'livekit-client';
 import { createLiveKitBus } from '../../../lib/livekit/livekit-bus';
-import { useContextKey } from '../RoomScopedProviders';
+import { useContextKey } from '@/components/RoomScopedProviders';
 import { useRealtimeSessionTranscript } from '@/hooks/use-realtime-session-transcript';
 import { supabase } from '@/lib/supabase';
 import { CanvasLiveKitContext } from '../livekit/livekit-room-connector';
