+++ src/components/ui/tldraw/hooks/useEditorReady.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { useEditor, Editor } from '@tldraw/tldraw';
+import { useEditor, Editor } from 'tldraw';
 
 interface EditorReadyResult {
   editor: Editor | null;
