+++ src/components/ui/tldraw/utils/canvas-selection-shared.ts
@@ -1,5 +1,5 @@
 import { toPlainText } from './rich-text';
-import type { Editor } from '@tldraw/tldraw';
+import type { Editor } from 'tldraw';
 
 export function resolveTargetShapeId(
   editor: Editor,
