+++ src/components/ui/tldraw/utils/pinned-shapes.ts
@@ -1,4 +1,4 @@
-import type { Editor } from '@tldraw/tldraw';
+import type { Editor } from 'tldraw';
 import type { customShape } from '../tldraw-canvas';
 
 type ViewportBounds = { width: number; height: number };
